+++ R/inclusion_prob.R
@@ -22,7 +22,7 @@
 #' `inclusion_prob()` returns a numeric vector of inclusion probabilities for
 #' each unit in the population.
 #'
-#' `becomes_ta()` returns an integer vector giving the sample size at which a
+#' `becomes_ta()` returns a numeric vector giving the sample size at which a
 #' unit enters the take-all stratum.
 #'
 #' @note